Seperate ApplicationDBContext
Remove Application DbContext in to own class
