+++ SwimmingAPI/SwimmingAPI.csproj
@@ -243,6 +243,7 @@
     <Compile Include="Migrations\Configuration.cs" />
     <Compile Include="Models\AccountBindingModels.cs" />
     <Compile Include="Models\AccountViewModels.cs" />
+    <Compile Include="Models\ApplicationDbContext.cs" />
     <Compile Include="Models\EventBindings.cs" />
     <Compile Include="Models\IdentityModels.cs" />
     <Compile Include="Models\MeetBindingModels.cs" />
